changed the emis_code ,student_id and admission number to biginteger for student

+++ oosc/oosc/students/models.py
@@ -5,15 +5,15 @@ from oosc.schools.models import Schools
 # Create your models here.
 class Students(models.Model):
     GENDERS = (('M', 'MALE'), ('F', 'FEMALE'))
-    student_id    = models.IntegerField(default=0,null=True,blank=True)
+    student_id    = models.BigIntegerField(default=0,null=True,blank=True)
     #school_id     = models.ForeignKey(Schools,on_delete = models.CASCADE)
-    emis_code     = models.IntegerField(default=0,null=True,blank=True)
+    emis_code     = models.BigIntegerField(default=0,null=True,blank=True)
     fstname  = models.CharField(max_length=200)
     midname = models.CharField(max_length=200,null=True,blank=True)
     lstname = models.CharField(max_length=200)
     date_of_birth = models.DateField(null=True,blank=True)
     date_enrolled=models.DateField(auto_created=True)
-    admission_no  = models.IntegerField(default=0,null=True,blank=True)
+    admission_no  = models.BigIntegerField(default=0,null=True,blank=True)
     class_id      = models.ForeignKey(Stream, on_delete = models.CASCADE) #shows the current class
     gender        = models.CharField(max_length=2,choices=GENDERS, default='ML')
     previous_class    = models.IntegerField(default=0,null=True,blank=True)
